refactor: 이벤트 생성 페이지 리팩토링 (#385)
* refactor: event name 관련 커스텀 훅 분리

* style: 필요하지 않은 children 제거

* feat: navigate replace true 조건으로 입력 중 상태로 되돌아갈 수 없도록 설정

* fix: 비밀번호 입력 창에서는 뒤로가기 할 때 이름으로 가도록 설정
